% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{buildVeloviz}
\alias{buildVeloviz}
\title{Main function}
\usage{
buildVeloviz(
  curr,
  proj,
  normalize.depth = TRUE,
  depth = 1e+06,
  use.ods.genes = TRUE,
  alpha = 0.05,
  pca = TRUE,
  center = TRUE,
  scale = TRUE,
  nPCs = 10,
  k = 10,
  similarity.threshold = 0,
  distance.weight = 1,
  weighted = FALSE,
  verbose = FALSE,
  details = FALSE,
  seed = 0
)
}
\description{
Main function
}
