% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectedNeighbors.R
\name{graphViz}
\alias{graphViz}
\title{Visualize as velocity informed force directed graph}
\usage{
graphViz(
  observed,
  projected,
  k,
  distance_metric = "L2",
  similarity_metric = "cosine",
  distance_weight = 1,
  distance_threshold = 1,
  similarity_threshold = -1,
  weighted = TRUE,
  remove_unconnected = TRUE,
  cell.colors = NA,
  title = NA,
  plot = TRUE,
  return_graph = FALSE
)
}
\description{
Visualize as velocity informed force directed graph
}
