% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MERFISH}
\alias{MERFISH}
\title{MERFISH scRNA-seq data}
\format{
list of 6 objects:
\describe{
\item{nuc}{matrix, 9050 genes x 645 cells of nuclear counts}
\item{cyto}{matrix, 9050 genes x 645 cells of cytoplasmic counts}
\item{col}{factor of cell colors corresponding to position in cell cycle}
\item{pcs}{matrix, 645 x 50 cell scores in 50 PCs}
\item{cell.dist}{dist, pairwise cell distances in PC space used to compute velocity}
\item{vel}{list, output from running velocyto using spliced, unspliced, and cell.dist}
}
}
\source{
\url{https://www.pnas.org/content/116/39/19490}
}
\usage{
MERFISH
}
\description{
MERFISH sequencing of U20S cells in culture, subsampled to 645 cells.
}
\keyword{datasets}
