% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeDepth}
\alias{normalizeDepth}
\title{Normalizes counts to CPM}
\usage{
normalizeDepth(counts, depthScale = 1e+06, verbose = TRUE)
}
\arguments{
\item{counts}{Read count matrix. The rows correspond to genes, columns
correspond to individual cells}

\item{depthScale}{Depth scaling. Using a million for CPM (default: 1e6)}

\item{verbose}{Boolean for verbosity setting (default: TRUE)}
}
\value{
a normalized matrix
}
\description{
Normalizes raw counts to counts per million
}
\examples{
vel <- pancreas$vel
curr <- vel$current

normalizeDepth(curr)

}
