% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreasWithGap}
\alias{pancreasWithGap}
\title{Pancreas scRNA-seq data missing intermediates}
\format{
list of 4 objects:
\describe{
\item{spliced}{matrix, 7192 genes x 660 cells of spliced counts}
\item{unspliced}{matrix, 7192 genes x 660 cells of unspliced counts}
\item{pcs}{matrix, 660 x 50 cell scores in 50 PCs}
\item{clusters}{factor of cell type annotations from scVelo}
}
}
\source{
\url{https://dev.biologists.org/content/146/12/dev173849.long}
}
\usage{
pancreasWithGap
}
\description{
Pancreatic endocrinogenesis scRNA-seq from Bastidas-Ponce et. al.,
Development 2019 accessed via scVelo package with missing Ngn3 high EP cells
and subsampled to 660 cells.
}
\keyword{datasets}
