% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotVeloviz}
\alias{plotVeloviz}
\title{Plot function}
\usage{
plotVeloviz(
  vig,
  layout.method = igraph::layout_with_fr,
  clusters = NA,
  cluster.method = igraph::cluster_louvain,
  col = NA,
  alpha = 0.05,
  verbose = TRUE,
  seed = 0
)
}
\description{
Plot function
}
