% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectedNeighbors.R
\name{projectedNeighbors}
\alias{projectedNeighbors}
\title{Get projected neighbors}
\usage{
projectedNeighbors(
  observed,
  projected,
  k,
  distance_metric = "L2",
  similarity_metric = "cosine",
  distance_weight = 1,
  distance_threshold = 1,
  similarity_threshold = -1
)
}
\description{
Get projected neighbors
}
