% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vel}
\alias{vel}
\title{MERFISH velocity subset}
\format{
list of 1:
\describe{
\item{vel}{velocity output containing current observed ("current") and predicted future ("projected") estimates}
}
}
\source{
\url{https://www.pnas.org/content/116/39/19490}
}
\usage{
vel
}
\description{
output of velocyto.R::gene.relative.velocity.estimates for 40 cell subset of MERFISH data. Used to run examples
}
\keyword{datasets}
