% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{veloviz}
\alias{veloviz}
\title{Main function}
\usage{
veloviz(
  mat,
  curr,
  proj,
  cell.cols = NA,
  center = TRUE,
  scale = FALSE,
  alpha = 0.05,
  nPCs = 10,
  k = 10,
  similarity.threshold = 0,
  distance.weight = 1,
  weighted = FALSE,
  cluster.method = igraph::cluster_louvain,
  layout.method = igraph::layout_with_fr,
  plot = TRUE,
  verbose = TRUE,
  seed = 0
)
}
\description{
Main function
}
