% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MERFISH}
\alias{MERFISH}
\title{MERFISH scRNA-seq data}
\format{
list of 4 objects:
\describe{
\item{nuc}{matrix, 9050 genes x 225 cells of nuclear counts}
\item{cyto}{matrix, 9050 genes x 225 cells of cytoplasmic counts}
\item{pcs}{matrix, 225 x 50 cell scores in 50 PCs}
\item{col}{factor of cell colors corresponding to position in cell cycle}
}
}
\source{
\url{https://www.pnas.org/content/116/39/19490}
}
\usage{
MERFISH
}
\description{
MERFISH sequencing of U20S cells in culture, subsampled to 225 cells.
}
\keyword{datasets}
