% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreasWithGap}
\alias{pancreasWithGap}
\title{Pancreas subsetted dataset}
\format{
Info
}
\source{
\url{link}
}
\usage{
pancreasWithGap
}
\description{
Pancreas subsetted dataset
}
\keyword{datasets}
