% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{reduceDimensions}
\alias{reduceDimensions}
\title{PCA}
\usage{
reduceDimensions(
  cpm,
  center = TRUE,
  scale = FALSE,
  use.ods.genes = TRUE,
  max.ods.genes = 2000,
  alpha = 0.05,
  nPCs = 50,
  verbose = TRUE,
  plot = FALSE,
  details = FALSE
)
}
\description{
PCA
}
